@@ -1,9 +1,19 @@
-'use strict';
-var fs = require('fs');
-var test = require('ava');
-var getEmails = require('./');
+import fs from 'fs';
+import test from 'ava';
+import m from './';
 
-test(function (t) {
-	t.plan(1);
-	t.assert(getEmails(fs.readFileSync('fixture.txt', 'utf8')).length === 5);
+// Fix Babel on Node.js 4
+const Set = global.Set;
+
+test(t => {
+	t.deepEqual(
+		m(fs.readFileSync('fixture.txt', 'utf8')),
+		new Set([
+			'sindresorhus@gmail.com',
+			'foo@bar.co.uk',
+			'unicorn@unicorn.unicorn',
+			'cake@cake.cake',
+			'pony@power.com'
+		])
+	);
 });
