@@ -1,11 +1,7 @@
 'use strict';
-var arrayUniq = require('array-uniq');
-var emailRegex = require('email-regex');
+const emailRegex = require('email-regex');
 
-module.exports = function (str) {
-	var emails = str.match(emailRegex());
-
-	return emails ? arrayUniq(emails.map(function (el) {
-		return el.trim();
-	})) : [];
+module.exports = input => {
+	const emails = input.match(emailRegex());
+	return new Set(emails ? emails.map(x => x.trim()) : []);
 };
